+++ include/H2DE/objects/H2DE_button_object.h
@@ -273,7 +273,6 @@ private:
 
     H2DE_TextObject* _textObject = H2DE_NULL_OBJECT;
     std::unordered_map<std::string, H2DE_Surface*> _surfaces = {};
-    H2DE_ButtonEventData _eventData;
 
     bool _disabled = false;