+++ src/objects/H2DE_button_object.cpp
@@ -1,17 +1,15 @@
 #include "H2DE/objects/H2DE_button_object.h"
 
 // INIT
-H2DE_ButtonObject::H2DE_ButtonObject(H2DE_Engine* e, const H2DE_ObjectData& od, const H2DE_ButtonObjectData& bod) : H2DE_Object(e, od), _buttonObjectData(bod), _eventData({ this, nullptr }) {
+H2DE_ButtonObject::H2DE_ButtonObject(H2DE_Engine* e, const H2DE_ObjectData& od, const H2DE_ButtonObjectData& bod) : H2DE_Object(e, od), _buttonObjectData(bod) {
+    _buttonObjectData.text.text = H2DE_TextObject::_getFormatedText(_buttonObjectData.text.text);
+
     _refreshSurfaceBuffers();
     _refreshMaxRadius();
 }
 
 // CLEANUP
 H2DE_ButtonObject::~H2DE_ButtonObject() {
-    if (_eventData.timeline != nullptr) {
-        _eventData.timeline->stop(false);
-    }
-
     if (_textObject != H2DE_NULL_OBJECT) {
         if (_engine->destroyObject(_textObject)) {
             _textObject = H2DE_NULL_OBJECT;
@@ -46,34 +44,34 @@ void H2DE_ButtonObject::_refreshMaxRadius() {
 
 void H2DE_ButtonObject::mouseDown() {
     if (_buttonObjectData.onMouseDown && !_disabled) {
-        _buttonObjectData.onMouseDown(_eventData);
+        _buttonObjectData.onMouseDown({ this });
     }
 }
 
 void H2DE_ButtonObject::mouseUp() {
     if (_buttonObjectData.onMouseUp && !_disabled) {
-        _buttonObjectData.onMouseUp(_eventData);
+        _buttonObjectData.onMouseUp({ this });
     }
 }
 
 void H2DE_ButtonObject::mouseHover() {
     if (_buttonObjectData.onHover && !_disabled) {
-        _buttonObjectData.onHover(_eventData);
+        _buttonObjectData.onHover({ this });
     }
 }
 
 void H2DE_ButtonObject::mouseBlur() {
     if (_buttonObjectData.onBlur && !_disabled) {
-        _buttonObjectData.onBlur(_eventData);
+        _buttonObjectData.onBlur({ this });
     }
 }
 
 // SETTER
 void H2DE_ButtonObject::setText(const std::string& text) {
-    if (text == _buttonObjectData.text.text) {
+    if (H2DE_TextObject::_getFormatedText(text) == _buttonObjectData.text.text) {
         return;
     }
 
-    _buttonObjectData.text.text = text;
+    _buttonObjectData.text.text = H2DE_TextObject::_getFormatedText(text);
     _refreshTextObject();
 }